// frontend/src/shared/api/schema-auth.d.ts
export interface components {
    schemas: {
        AccessTokenResponse: {
            readonly tokenType?: string | null;
            accessToken: string | null;
            /** Format: int64 */
            expiresIn: number;
            refreshToken: string | null;
        };
        ForgotPasswordRequest: {
            email: string | null;
        };
        HttpValidationProblemDetails: {
            type?: string | null;
            title?: string | null;
            /** Format: int32 */
            status?: number | null;
            detail?: string | null;
            instance?: string | null;
            errors?: {
                [key: string]: string[];
            } | null;
        } & {
            [key: string]: unknown;
        };
        LoginDto: {
            email: string | null;
            password: string | null;
            twoFactorCode?: string | null;
            twoFactorRecoveryCode?: string | null;
            useCookies?: boolean | null;
            rememberMe?: boolean | null;
        };
        ProfileOut: {
            /** Format: uuid */
            id?: string;
            login: string | null;
            firstName?: string | null;
            lastName?: string | null;
            description?: string | null;
            companyWebsite?: string | null;
        };
        RefreshRequest: {
            refreshToken: string | null;
        };
        RegisterRequest: {
            email: string | null;
            password: string | null;
        };
        ResendConfirmationEmailRequest: {
            email: string | null;
        };
        ResetPasswordRequest: {
            email: string | null;
            resetCode: string | null;
            newPassword: string | null;
        };
        TwoFactorRequest: {
            enable?: boolean | null;
            twoFactorCode?: string | null;
            resetSharedKey?: boolean;
            resetRecoveryCodes?: boolean;
            forgetMachine?: boolean;
        };
        TwoFactorResponse: {
            sharedKey: string | null;
            /** Format: int32 */
            recoveryCodesLeft: number;
            recoveryCodes?: string[] | null;
            isTwoFactorEnabled: boolean;
            isMachineRemembered: boolean;
        };
        UpdateProfileDto: {
            firstName?: string | null;
            lastName?: string | null;
            description?: string | null;
            companyWebsite?: string | null;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}